+++ src/base/ConfigMgr.cpp
@@ -170,13 +170,13 @@ int ConfigMgr::getIntOption(const std::string& sSubsys,
 }
 
 void ConfigMgr::getGammaOption(const std::string& sSubsys, 
-            const std::string& sName, double* Val) const
+            const std::string& sName, float* Val) const
 {
     const string * psOption = getOption(sSubsys, sName);
     if (psOption == 0) {
         return;
     }
-    int rc = sscanf(psOption->c_str(), "%lf,%lf,%lf", Val, Val+1, Val+2);
+    int rc = sscanf(psOption->c_str(), "%f,%f,%f", Val, Val+1, Val+2);
     if (rc < 3) {
         AVG_TRACE(Logger::ERROR,
                 m_sFName << ": Unrecognized value for option "<<sName<<": "
