+++ src/base/ConfigMgr.h
@@ -59,7 +59,7 @@ public:
     int getIntOption(const std::string& sSubsys, 
             const std::string& sName, int Default) const;
     void getGammaOption(const std::string& sSubsys, 
-            const std::string& sName, double* Val) const;
+            const std::string& sName, float* Val) const;
 
     void dump() const;
