+++ src/base/CubicSpline.cpp
@@ -28,15 +28,15 @@ using namespace std;
 
 namespace avg {
 
-CubicSpline::CubicSpline(const vector<double>& x, const vector<double>& y)
+CubicSpline::CubicSpline(const vector<float>& x, const vector<float>& y)
     : m_X(x),
       m_Y(y)
 {
     AVG_ASSERT(x.size() == y.size());
     // Add fake points before the first and after the last point so all derivatives
     // are defined.
-    double edgeX = 2*m_X[0]-m_X[1];
-    double edgeY = 2*m_Y[0]-m_Y[1];
+    float edgeX = 2*m_X[0]-m_X[1];
+    float edgeY = 2*m_Y[0]-m_Y[1];
     m_X.insert(m_X.begin(), edgeX);
     m_Y.insert(m_Y.begin(), edgeY);
 
@@ -51,19 +51,19 @@ CubicSpline::~CubicSpline()
 {
 }
 
-double normedInterpolate(double y0, double y1, double y2, double y3, double mu)
+float normedInterpolate(float y0, float y1, float y2, float y3, float mu)
 {
-   double mu2 = mu*mu;
-   double a0 = y3 - y2 - y0 + y1;
-   double a1 = y0 - y1 - a0;
-   double a2 = y2 - y0;
-   double a3 = y1;
+   float mu2 = mu*mu;
+   float a0 = y3 - y2 - y0 + y1;
+   float a1 = y0 - y1 - a0;
+   float a2 = y2 - y0;
+   float a3 = y1;
 
    return(a0*mu*mu2+a1*mu2+a2*mu+a3);
 }
 
 
-double CubicSpline::interpolate(double orig)
+float CubicSpline::interpolate(float orig)
 {
     unsigned i = 0;
     if (m_X[m_X.size()-1] <= orig) {
@@ -74,14 +74,14 @@ double CubicSpline::interpolate(double orig)
         }
     }
     if (i < 2) {
-        double dxdy = (m_X[1]-m_X[0])/(m_Y[1]-m_Y[0]);
+        float dxdy = (m_X[1]-m_X[0])/(m_Y[1]-m_Y[0]);
         return m_Y[1]+(orig-m_X[1])/dxdy;
     } else  if (i > m_X.size()-2) {
         unsigned len = m_X.size();
-        double dxdy = (m_X[len-1]-m_X[len-2])/(m_Y[len-1]-m_Y[len-2]);
+        float dxdy = (m_X[len-1]-m_X[len-2])/(m_Y[len-1]-m_Y[len-2]);
         return m_Y[len-2]+(orig-m_X[len-2])/dxdy;
     } else {
-        double ratio = (orig-m_X[i-1])/(m_X[i]-m_X[i-1]);
+        float ratio = (orig-m_X[i-1])/(m_X[i]-m_X[i-1]);
         return normedInterpolate(m_Y[i-2], m_Y[i-1], m_Y[i], m_Y[i+1], ratio);
     }
 }
