+++ src/graphics/Filtercolorize.cpp
@@ -28,45 +28,45 @@
 namespace avg {
 
 // From Foley, van Dam p. 596 incl. addendum fixes.
-unsigned char hls_value(double n1, double n2, double hue)
+unsigned char hls_value(float n1, float n2, float hue)
 {
   if (hue>360)
     hue-=360;
   if (hue<0)   
     hue+=360;
   
-  double rv;
+  float rv;
   if (hue<60)  
-    rv = n1+(n2-n1)*hue/60.0;
+    rv = n1+(n2-n1)*hue/60;
   else if (hue<180) 
     rv = n2;
   else if (hue<240) 
-    rv = n1+(n2-n1)*(240.0-hue)/60.0;
+    rv = n1+(n2-n1)*(240-hue)/60;
   else 
     rv = n1;
 
   return (unsigned char)(rv*255);
 }
 
-Pixel24 hls2rgb (double h, double l, double s)
+Pixel24 hls2rgb (float h, float l, float s)
 {
-    double m1, m2;
+    float m1, m2;
     l /= 255;
     s /= 100;
     // Warning: Foley, van Dam has a typo on the next line!
-    m2 = (l<=0.5)?(l*(1.0+s)):(l+s-l*s);
-    m1 = 2.0*l-m2;
-    if (s<0.001) {
+    m2 = (l<=0.5f)?(l*(1.0f+s)):(l+s-l*s);
+    m1 = 2.0f*l-m2;
+    if (s<0.001f) {
         return Pixel24((unsigned char)(l*255), (unsigned char)(l*255), 
                 (unsigned char)(l*255));
     } else {
-        return Pixel24(hls_value(m1,m2,h+120.0),
+        return Pixel24(hls_value(m1,m2,h+120),
                 hls_value(m1,m2,h),
-                hls_value(m1,m2,h-120.0));
+                hls_value(m1,m2,h-120));
     }
 }
 
-FilterColorize::FilterColorize(double hue, double saturation)
+FilterColorize::FilterColorize(float hue, float saturation)
   : m_Hue(hue),
     m_Saturation(saturation)
 {
@@ -82,7 +82,7 @@ void FilterColorize::applyInPlace(BitmapPtr pBmp)
     BitmapPtr pTempBmp (FilterGrayscale().apply(pBmp));
     Pixel24 colorTable[256];
     for (int i=0; i<256; i++) {
-        colorTable[i] = hls2rgb(m_Hue, i, m_Saturation);
+        colorTable[i] = hls2rgb(m_Hue, float(i), m_Saturation);
     }
 
     unsigned char * pSrcLine = pTempBmp->getPixels();
