+++ src/imaging/CMUCameraUtils.h
@@ -36,7 +36,7 @@ namespace avg {
 
 void getVideoFormatAndMode(IntPoint& Size, PixelFormat pf,
         unsigned long* pVideoFormat, unsigned long* pVideoMode);
-unsigned long getFrameRateConst(double FrameRate);
+unsigned long getFrameRateConst(float FrameRate);
 float getFrameRateFloat(unsigned long FrameRate);
 CAMERA_FEATURE getFeatureID(CameraFeature Feature);
 void getImageSizeAndPF(unsigned long videoFormat, unsigned long videoMode,
